requirements.txt
asgiref==3.6.0
autopep8==2.0.2
bcrypt==4.0.1
boto3==1.25.4
botocore==1.28.4
cachetools==5.2.0
certifi==2022.12.7
cffi
charset-normalizer==2.1.1
colorama==0.4.6
cryptography==39.0.0
defusedxml==0.7.1
Django==4.1.5
django-4-jet==1.0.9
django-allauth==0.52.0
django-cors-headers==3.13.0
django-dotenv==1.4.2
django-environ==0.9.0
django-storages==1.13.1
djangorestframework==3.14.0
google-api-core==2.10.2
google-api-python-client==2.65.0
google-auth==2.13.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.7.0
googleapis-common-protos==1.56.4
google-cloud-storage==2.10.0
httplib2==0.21.0
idna==3.4
importlib-metadata==5.0.0
iniconfig==2.0.0
jmespath==1.0.1
Markdown==3.4.1
numpy
oauthlib==3.2.2
packaging==24.2
pandas
pluggy==1.5.0
protobuf==4.21.9
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.10.0
pycparser==2.21
PyJWT==2.6.0
PyMySQL==1.0.2
pyparsing==3.0.9
pytest==8.3.4
pytest-django==4.9.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2022.2.1
qrcode==7.3.1
razorpay==1.3.0
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
types-cryptography==3.3.23.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==1.26.13
zipp==3.10.0
gunicorn==21.2.0
Pillow==9.5.0
pycryptodome==3.18.0
setuptools==67.8.0

whitenoise==6.6.0
gspread==5.7.2
